feat: CourseSummarySerializer 구현
